Update latest BashPass version to '3.1'
This commit updates the version of BashPass to 3.1 and changes the package names in the build-bashpass.yml file to match the new version. The bashpass.spec file has also been updated to reflect the new version. This commit ensures that the latest package is uploaded to the PackageCloud repositories for Fedora 36 and 37.